extract Java7 methods into base
